export buildEmailLayoutDirBundle helper
