+++ src/isomorphic.ts
@@ -1,0 +1,4 @@
+/*
+ * IMPORTANT: You must only expose exports from isomorphic modules.
+ */
+export * as marshal from "./lib/marshal/index.isomorphic";
