+++ src/lib/marshal/email-layout/index.ts
@@ -1,4 +1,5 @@
 export * from "./helpers";
+export * from "./processor.isomorphic";
 export * from "./reader";
 export * from "./types";
 export * from "./writer";
