+++ src/lib/marshal/email-layout/reader.ts
@@ -22,9 +22,9 @@ import { EmailLayoutDirContext } from "@/lib/run-context";
 import {
   EmailLayoutCommandTarget,
   isEmailLayoutDir,
-  LAYOUT_JSON,
   lsEmailLayoutJson,
 } from "./helpers";
+import { LAYOUT_JSON } from "./processor.isomorphic";
 
 type JoinExtractedFilesResult = [AnyObj, JsonDataError[]];
