+++ src/lib/marshal/index.isomorphic.ts
@@ -1,0 +1,5 @@
+/*
+ * IMPORTANT: You must only expose exports from isomorphic modules.
+ */
+export { buildWorkflowDirBundle } from "./workflow/processor.isomorphic";
+export { buildEmailLayoutDirBundle } from "./email-layout/processor.isomorphic";
